Rename engine package pcgen.core.doomsdaybook to pcgen.core.namegen
The 'doomsdaybook' name is a 2003 artifact of the third-party product
PCGen adopted the engine from. Rename clarifies the package's purpose.
