+++ code/src/java/pcgen/core/namegen/Operation.java
@@ -16,7 +16,7 @@
  *  License along with this library; if not, write to the Free Software
  *  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  */
-package pcgen.core.doomsdaybook;
+package pcgen.core.namegen;
 
 /**
  * {@code Operation} encapsulates an action that can be performed
